working but awful
