+++ src/main/java/BootApp.java
@@ -1,20 +1,18 @@
-import app.TriggerEndpointConfiguration;
-import app.UserConfiguration;
-import app.UserConfigurationBuilder;
+import configuration.TriggerEndpointConfiguration;
+import configuration.UserConfiguration;
 import com.google.common.collect.ImmutableList;
 
 public class BootApp
 {
     public static void main(String[] args)
     {
-        final UserConfiguration userConfiguration = UserConfigurationBuilder.builder()
-                .withPort(8090)
-                .withBrokerDestinationEndpoint("/topic")
-                .withStompConnectionEndpoint("/livescore-websocket")
-                .withTriggerEndpointConfigurations(ImmutableList.of(
-                        new TriggerEndpointConfiguration("/app", ImmutableList.of("/topic/myscores"))
-                ))
-                .build();
+        final UserConfiguration userConfiguration = UserConfiguration.tryCreatingSingleton(
+                8090,
+                "/livescore-websocket",
+                "/topic",
+                ImmutableList.of(
+                        new TriggerEndpointConfiguration("/app", ImmutableList.of("/topic/myscores")))
+        );
 
         PushServer.run(userConfiguration, args);
     }
