+++ src/main/java/PushServer.java
@@ -1,6 +1,5 @@
-import app.UserConfiguration;
 import configuration.NotificationSocketConfig;
-import org.springframework.beans.factory.config.ConfigurableListableBeanFactory;
+import configuration.UserConfiguration;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
@@ -8,14 +7,12 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@ComponentScan({"hello"})
+@ComponentScan({"controllers"})
 @Import(NotificationSocketConfig.class)
 public class PushServer
 {
     public static void run(UserConfiguration userConfiguration, String... args)
     {
         final ConfigurableApplicationContext context = SpringApplication.run(new Object[]{PushServer.class}, args);
-        final ConfigurableListableBeanFactory beanFactory = context.getBeanFactory();
-        beanFactory.registerSingleton(UserConfiguration.class.getName(), userConfiguration);
     }
 }
